refactor(#3479): extract catch-up subsystem from discord/mod.rs to catch_up.rs
Behavior-preserving verbatim move of the catch-up (restart-gap message
recovery) cluster + its tests; discord/mod.rs drops 4979 -> 4243 prod LoC
(catch_up.rs 730 prod LoC, under the giant-file threshold) and the frozen
giant baseline is ratcheted down to 4243 to lock in the shrink.
