+++ scripts/audit_maintainability_giant_baseline.toml
@@ -139,7 +139,11 @@
 # #3038 S5: lowered 4967 -> 4944 (-23) as the cluster-G runtime HTTP cache
 # fields plus the `serenity_http_or_token_fallback` accessor moved into
 # `shared_state::RuntimeHttpCache`. Locks in the shrink win (ratchet down only).
-"src/services/discord/mod.rs" = 4979
+# #3479 item-2 (behavior-preserving extraction): the catch-up subsystem
+# (restart-gap message recovery — classify/scan/enqueue + its tests) moved
+# verbatim to the capped sibling `catch_up.rs`, lowering the frozen baseline
+# 4979 -> 4243 (-736). Locks in the shrink win (#3028 ratchet down only).
+"src/services/discord/mod.rs" = 4243
 # #3305: the local-only pass-through lifecycle-skip (kind hoist + note-only return
 # branch + idle-loop skip) is offset by comment dedup in the same root, lowering the
 # frozen baseline 5438 -> 5429 (-9). Locks in the shrink win (#3028 ratchet).
